+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ This document records the main changes to the tree code.
 - Updated the path syntax for special functions to "@[function_name]|" to resolve issues with inclusion next to other characters
 - Added new sdss5 paths for apogee
 - Changed sdss5 paths for apogee from sdss4 to handle ``healpix`` grouping and simplified kwargs with special functions
+- Updating the code to write out old sdss_paths.ini file.
 
 3.0.6 (2020-09-07)
 ------------------